backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
mysql-connector-python==8.2.0
pymysql==1.1.0
sqlalchemy==2.0.23
redis==5.0.1
bcrypt==4.1.1
numpy==1.24.4
scipy==1.11.3
pandas==2.1.3
openpyxl==3.1.2
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
albumentations==1.3.0
insightface==0.7.3
onnxruntime-gpu==1.16.3
faiss-gpu==1.7.4
psycopg2-binary
python-dotenv
